[Refactor] useImageUploadWidget composable
